Seperate Language chart
The chart of languages has been split up in a different
component. This reduced the number of state refreshes.
