+++ src/components/atoms/index.js
@@ -1,11 +1,13 @@
 import ScrollToTop from "./ScrollToTop";
 import Calendar3D from "./Calendar3D";
 import Calendar2D from "./Calendar2D";
+import LanguageDoughnut from "./LanguageDoughnut";
 
 export {
   ScrollToTop,
   Calendar3D,
   Calendar2D,
+  LanguageDoughnut,
 }
 
 /**
